requirements.txt
azure-common==1.1.26
azure-core==1.9.0
azure-identity==1.5.0
azure-keyvault==4.1.0
azure-keyvault-certificates==4.2.1
azure-keyvault-keys==4.3.1
azure-keyvault-secrets==4.2.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
Django==2.2
django-crispy-forms==1.9.2
django-localflavor==3.0.1
django-storages==1.11
idna==2.10
isodate==0.6.0
msal==1.7.0
msal-extensions==0.3.0
msrest==0.6.19
oauthlib==3.1.0
Pillow==8.0.0
portalocker==1.7.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-stdnum==1.14
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
zipcodes==1.1.2
